// src/app/core/resources/dto/transaction.ts
/**
 * BWZ-Rappi - Bank of Rapperswil Backend
 * Das Backend für die Bank of Rapperswil stellt eine vereinfachte API für das Management von Finanz-Transaktionen als Übungsplattform bereit.
 *
 * OpenAPI spec version: v0.51.1
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Enthält die Information für die Überweisung eines Betrags von Konto A nach B (target).
 */
export interface Transaction { 
    /**
     * Spezifiziert den Account für die Gutschrift.
     */
    readonly target: string;
    /**
     * Gibt den zu überweisenden Betrag an.
     */
    readonly amount: number;
    /**
     * Optional, spezifiziert die Ausgabenkategorie für die Answertung. Initial werden die folgenden Kategorien gefüllt: 'living', 'food', 'furniture', 'mobility', 'beauty_health', 'media', 'electronics'
     */
    readonly category?: string;
}