// src/app/core/resources/dto/transaction-confirmation.ts
/**
 * BWZ-Rappi - Bank of Rapperswil Backend
 * Das Backend für die Bank of Rapperswil stellt eine vereinfachte API für das Management von Finanz-Transaktionen als Übungsplattform bereit.
 *
 * OpenAPI spec version: v0.51.1
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Enthält die Bestätigungsinformation einer getätigten Überweisung.
 */
export interface TransactionConfirmation { 
    /**
     * Spezifiziert den Account der Lastschrift.
     */
    readonly source: string;
    /**
     * Spezifiziert den Account für die Gutschrift.
     */
    readonly target: string;
    /**
     * Enthält den überwiesenen Betrag.
     */
    readonly amount: number;
    /**
     * Optional, spezifiziert die Ausgabenkategorie für die Answertung.
     */
    readonly category?: string;
    /**
     * Enthält den neuen Saldo auf dem Konto an.
     */
    readonly newBalance: number;
    /**
     * Enthält den Zeitpunkt der Transaktion im ISO-String Format.
     */
    readonly date: string;
}